refactor(client): dedup extractMotionMeta into shared motion utility
extractMotionMeta was duplicated verbatim between previewApi.ts (1
call site) and useGenerationActions.ts (4 call sites) — they had to
stay in lockstep for telemetry. Extract into client/src/utils/motion.ts
so there is one canonical implementation.

Also restore the defensive result.data null guard in motionApi that
the discriminated-union narrowing made type-redundant but runtime-
useful — protects against a server returning success: true without
data.
